final_pjt_back/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
idna==3.4
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
